+++ Changelog.md
@@ -1,3 +1,4 @@
 
 July 24 2020
 - Fix the bug in parsing pong message issue -> https://github.com/ant-media/Ant-Media-Server/issues/2362
+- Make Android SDK DataChannel compatible with Web. MainActivity can be used to send text messages through data channel to web https://github.com/ant-media/Ant-Media-Server/issues/2354